fix multiple gitlab source in build will cause token refresh error
